Сводный отчет за неделю (за месяц) по звонкам на e-mail

+++ trunk/kommunikator/reports/ats_report.sh
@@ -1,7 +1,17 @@
 #/usr/bin/sh
 tmp=/tmp
+p=/root/report
+sql=$p/ats_report$1.sql
+
+if [ ! -f $sql ]
+then
+   echo "File $sql does not exist."
+   exit 1;
+fi
+#exit 0;
 
 rm -f $tmp/ats_report.tmp
-mysql -uroot -proot < ./ats_report.sql
-iconv -t WINDOWS-1251 -f UTF-8 $tmp/ats_report.tmp > ./ats_report.csv
-./send_mail.sh
+mysql -uroot -proot < $sql
+iconv -t WINDOWS-1251 -f UTF-8 $tmp/ats_report.tmp > $p/ats_report.csv
+$p/send_mail.sh
+
